Modifications to the tree and task classes, updated prgm_test script
Added and updated several methods in the tree and task classes. The
prgm_test script was updated and currently works, however, it does not
yet test all the functionality of the program, task, and tree classes.
Additionally, there is no documentation on any of these classes.
Estimated that the class is 70% functional, but only 50% complete (due
to documentation).