Additional updates to tree & task classes
Cleared up several bugs with promote & move methods. Changed removechild
object to destroychild, and added a new removechild method. Moved both
of these methods from tree to task class. Modified the prgm_test script
to improve how it tests the move, promote, and disp functions. Added a
"ismember" method override to the tree class, this function takes in 2
lists of trees and checks if the elements in the first list are members
of the second list.

The tree & task classes are basically 100% implemented, just need to do
some thinking if there is any additional functionality that the classes
need. I also need to fully document these classes.