Documentation & findchild method written
This update added a bunch of documentation to the tree class.
Additionally the findchild method in the tree class was written to find
all items within a tree that match given criteria. This can replace the
findleaves function, but I haven't done that yet. I also haven't checked
if the findchild method works yet. Documentation ~40% done.